slow down Menu StartUp
When returning from game to menu, there was a problem when pressing the right side of the Menu button: the game would quit.

The problem seems to be that the menu gets loaded so fast that you are already pressing the EXIT button of themenu screen before you can even see it.

Makeshift solution: add a 0.1s delay in the Menu-function before entering the Menu-render-loop